--- a/lib/gateway.js
+++ b/lib/gateway.js
@@ -138,6 +138,8 @@ function defineResolvers (schema, typeToServiceMap, serviceMap, typeFieldsToServ
                 })
               }
             }
+          } else {
+            field.resolve = (parent, args, context, info) => parent && parent[info.path.key]
           }
         } else if (typeFieldsToService[`${type}-${fieldName}`]) {
           const service = serviceMap[typeFieldsToService[`${type}-${fieldName}`]]
@@ -155,6 +157,8 @@ function defineResolvers (schema, typeToServiceMap, serviceMap, typeFieldsToServ
               transformData: response => response.json.data._entities[0][fieldName]
             })
           }
+        } else {
+          field.resolve = (parent, args, context, info) => parent && parent[info.path.key]
         }
       }
     }