Fixes: https://github.com/mercurius-js/mercurius/issues/428

I have added tests for both gateway and non-gateway modes of operation. Next, I'll run the benchmarks and update here to make sure that this is the right approach.