@jonnydgreen 2021-03-25
Just ran some benchmarks and all looks good I think!

### Master
```
Running 60s test @ http://localhost:4000/graphql
100 connections
50 workers

┌─────────┬───────┬───────┬───────┬───────┬──────────┬─────────┬────────┐
│ Stat    │ 2.5%  │ 50%   │ 97.5% │ 99%   │ Avg      │ Stdev   │ Max    │
├─────────┼───────┼───────┼───────┼───────┼──────────┼─────────┼────────┤
│ Latency │ 14 ms │ 20 ms │ 30 ms │ 36 ms │ 20.27 ms │ 6.91 ms │ 640 ms │
└─────────┴───────┴───────┴───────┴───────┴──────────┴─────────┴────────┘
┌───────────┬────────┬────────┬────────┬─────────┬─────────┬────────┬────────┐
│ Stat      │ 1%     │ 2.5%   │ 50%    │ 97.5%   │ Avg     │ Stdev  │ Min    │
├───────────┼────────┼────────┼────────┼─────────┼─────────┼────────┼────────┤
│ Req/Sec   │ 1772   │ 3155   │ 4807   │ 5491    │ 4815.14 │ 590.57 │ 1772   │
├───────────┼────────┼────────┼────────┼─────────┼─────────┼────────┼────────┤
│ Bytes/Sec │ 362 kB │ 644 kB │ 980 kB │ 1.12 MB │ 982 kB  │ 120 kB │ 361 kB │
└───────────┴────────┴────────┴────────┴─────────┴─────────┴────────┴────────┘
```

### This PR

```
Running 60s test @ http://localhost:4000/graphql
100 connections
50 workers

┌─────────┬───────┬───────┬───────┬───────┬──────────┬─────────┬────────┐
│ Stat    │ 2.5%  │ 50%   │ 97.5% │ 99%   │ Avg      │ Stdev   │ Max    │
├─────────┼───────┼───────┼───────┼───────┼──────────┼─────────┼────────┤
│ Latency │ 14 ms │ 20 ms │ 30 ms │ 36 ms │ 20.18 ms │ 6.66 ms │ 592 ms │
└─────────┴───────┴───────┴───────┴───────┴──────────┴─────────┴────────┘
┌───────────┬────────┬────────┬────────┬─────────┬─────────┬────────┬────────┐
│ Stat      │ 1%     │ 2.5%   │ 50%    │ 97.5%   │ Avg     │ Stdev  │ Min    │
├───────────┼────────┼────────┼────────┼─────────┼─────────┼────────┼────────┤
│ Req/Sec   │ 1747   │ 3235   │ 4875   │ 5547    │ 4834.24 │ 596.51 │ 1747   │
├───────────┼────────┼────────┼────────┼─────────┼─────────┼────────┼────────┤
│ Bytes/Sec │ 357 kB │ 660 kB │ 994 kB │ 1.13 MB │ 986 kB  │ 122 kB │ 356 kB │
└───────────┴────────┴────────┴────────┴─────────┴─────────┴────────┴────────┘
```